api/requirements.txt
tensorflow==2.16.1
keras==3.3.3
fastapi
uvicorn
python-multipart
pillow
numpy<2.0
